[SPARK-32194][PYTHON] Add more refined python exceptions
