+++ python/pyspark/conf.py
@@ -157,7 +157,7 @@ class SparkConf(object):
     def setExecutorEnv(self, key=None, value=None, pairs=None):
         """Set an environment variable to be passed to executors."""
         if (key is not None and pairs is not None) or (key is None and pairs is None):
-            raise Exception("Either pass one key-value pair or a list of pairs")
+            raise ValueError("Either pass one key-value pair or a list of pairs")
         elif key is not None:
             self.set("spark.executorEnv." + key, value)
         elif pairs is not None:
