+++ python/pyspark/sql/tests/test_pandas_udf.py
@@ -19,7 +19,7 @@ import unittest
 
 from pyspark.sql.functions import udf, pandas_udf, PandasUDFType
 from pyspark.sql.types import DoubleType, StructType, StructField, LongType
-from pyspark.sql.utils import ParseException, PythonException
+from pyspark.sql.exceptions import ParseException, PythonException
 from pyspark.rdd import PythonEvalType
 from pyspark.testing.sqlutils import ReusedSQLTestCase, have_pandas, have_pyarrow, \
     pandas_requirement_message, pyarrow_requirement_message
